Изменен алгоритм хэширования на xxh3
